+++ examples/MyExampleEntity/MyCustomEntity.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+namespace examples\MyExampleEntity;
+
+use Vin\ShopwareSdk\Data\Entity\Entity;
+
+class MyCustomEntity extends Entity
+{
+}