+++ src/Data/Custom/CustomEntity.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+namespace Vin\ShopwareSdk\Data\Custom;
+
+use Vin\ShopwareSdk\Data\Entity\Entity;
+
+/**
+ * This class is generated dynamically following SW entities schema
+ */
+class CustomEntity extends Entity
+{
+}